single post data displayed

+++ controllers/api/comment-routes.js
@@ -25,6 +25,7 @@ router.post('/', (req, res) => {
 });
 
 router.delete('/:id', (req, res) => {
+  if(req.session){
     Comment.destroy({
       where: {
         id: req.params.id
@@ -41,6 +42,7 @@ router.delete('/:id', (req, res) => {
         console.log(err);
         res.status(500).json(err);
       });
+    }
   });
 
 module.exports = router;
